feat: #12 onReceive를 통해 playbackState receive onReceive를 통해 playbackstate 상태를 구독, 이를 player에 적용
