+++ Music-Player/Music-Player/View/ContentView.swift
@@ -20,72 +20,3 @@ struct ContentView: View {
         MiniPlayerView(player: $player, isPlaying: $isPlaying)
     }
 }
-
-class MiniPlayerViewModel: ObservableObject {
-    
-}
-
-struct MiniPlayerView: View {
-    
-    @ObservedObject var playerViewModel = MiniPlayerViewModel()
-    @Binding var player: MPMusicPlayerController
-    @Binding var isPlaying: Bool
-    @State var refreshView: Bool = false
-    
-    var body: some View {
-        HStack {
-            Button {
-                switch player.playbackState {
-                case .playing:
-                    print("pause song")
-                    pauseSong()
-                    break
-                default:
-                    print("play song")
-                    playSong()
-                }
-            } label: {
-                isPlaying ? Image(systemName: "pause.fill") : Image(systemName: "play.fill")
-            }
-            .font(.title)
-            Spacer()
-            VStack {
-                Text(player.nowPlayingItem?.title ?? undefinedString)
-                HStack {
-                    Text(player.nowPlayingItem?.artist ?? undefinedString)
-                        .lineLimit(1)
-                    Text(" ")
-                    Text(player.nowPlayingItem?.albumTitle ?? undefinedString)
-                        .lineLimit(1)
-                }
-            }
-            Spacer()
-            Image(uiImage: UIImage())
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(maxWidth: 50, maxHeight: 50)
-                .background(Color.gray)
-        }
-        .frame(height: 50)
-        .padding(EdgeInsets(top: 5, leading: 50, bottom: 5, trailing: 50))
-        .onAppear {
-            switch player.playbackState {
-            case .playing:
-                isPlaying = true
-                break
-            default:
-                isPlaying = false
-            }
-        }
-    }
-    
-    private func playSong() {
-        player.play()
-        isPlaying = true
-    }
-    
-    private func pauseSong() {
-        player.pause()
-        isPlaying = false
-    }
-}
